backend/requirements.txt
Django==4.2.8
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
sqlparse==0.5.0
typing-extensions==4.11.0
django-sslserver==0.22
pyotp==2.9.0
django-ses==4.0.0
qrcode==7.4.2
requests==2.31.0
requests-oauthlib==2.0.0
django-crontab==0.7.1
asgiref==3.7.2
Pillow==10.3.0
django-braces==1.15.0
django-embed-video==1.4.9
pymemcache==4.0.0
django-debug-toolbar==4.3.0
redis==5.0.4
channels[daphne]==4.1.0
channels-redis==4.2.0